pomelo-hw/requirements.txt
flask
requests
pymysql
Authlib
pyjwt
python-dotenv
